Moving tensorlfow_cloud to core
