+++ python/tensorflow_cloud/core/containerize.py
