+++ python/tensorflow_cloud/core/validate.py
