+++ python/tensorflow_cloud/tests/tensorflow_cloud/containerize_test.py
@@ -17,8 +17,8 @@ import os
 import tarfile
 import unittest
 
-from tensorflow_cloud import containerize
-from tensorflow_cloud import machine_config
+from tensorflow_cloud.core import containerize
+from tensorflow_cloud.core import machine_config
 
 from mock import call, patch
 
@@ -355,8 +355,8 @@ class TestContainerize(unittest.TestCase):
         os.remove(req_file)
         self.cleanup(lcb.docker_file_path)
 
-    @patch("tensorflow_cloud.containerize.logger")
-    @patch("tensorflow_cloud.containerize.APIClient")
+    @patch("tensorflow_cloud.core.containerize.logger")
+    @patch("tensorflow_cloud.core.containerize.APIClient")
     def test_get_docker_image(self, MockAPIClient, MockLogger):
         self.setup()
         mock_registry = "gcr.io/my-project"
