+++ python/tensorflow_cloud/tests/tensorflow_cloud/deploy_test.py
@@ -21,8 +21,8 @@ import sys
 import tarfile
 import unittest
 
-from tensorflow_cloud import deploy
-from tensorflow_cloud import machine_config
+from tensorflow_cloud.core import deploy
+from tensorflow_cloud.core import machine_config
 
 from mock import call, patch
 
@@ -75,7 +75,7 @@ class TestDeploy(unittest.TestCase):
         deploy.gcp.get_project_name = _mock_get_project_name
 
     @patch("sys.stdout", new_callable=io.StringIO)
-    @patch("tensorflow_cloud.deploy.discovery")
+    @patch("tensorflow_cloud.core.deploy.discovery")
     def test_deploy_job(self, MockDiscovery, MockStdOut):
         self.setup(MockDiscovery)
 
@@ -124,7 +124,7 @@ class TestDeploy(unittest.TestCase):
             ),
         )
 
-    @patch("tensorflow_cloud.deploy.discovery")
+    @patch("tensorflow_cloud.core.deploy.discovery")
     def test_request_dict_without_workers(self, MockDiscovery):
         self.setup(MockDiscovery)
         worker_count = 0
@@ -156,7 +156,7 @@ class TestDeploy(unittest.TestCase):
             },
         )
 
-    @patch("tensorflow_cloud.deploy.discovery")
+    @patch("tensorflow_cloud.core.deploy.discovery")
     def test_request_dict_without_user_args(self, MockDiscovery):
         self.setup(MockDiscovery)
         job_name = deploy.deploy_job(
@@ -184,7 +184,7 @@ class TestDeploy(unittest.TestCase):
             },
         )
 
-    @patch("tensorflow_cloud.deploy.discovery")
+    @patch("tensorflow_cloud.core.deploy.discovery")
     def test_request_dict_with_TPU_worker(self, MockDiscovery):
         self.setup(MockDiscovery)
         chief_config = machine_config.COMMON_MACHINE_CONFIGS["CPU"]
